changed required issues
