+++ src/darkTheme.tsx
@@ -1,6 +1,6 @@
 export const darkTheme = {
 	breakpoints: {
-		medium: "736px",
+		medium: "1040px",
 		large: "2000px",
 	},
 	colors: {
